refactor: 에러메시지 co-location 으로 구조 변경 #265
