+++ frontend/src/pages/mentoringCreate/constants/introduce.ts
@@ -1,3 +1,7 @@
 export const INTRODUCE = {
   MAX: 50,
 };
+
+export const ERROR_MESSAGE = {
+  INTRODUCE_MAX_LENGTH: `소개는 최대 ${INTRODUCE.MAX}자까지 입력할 수 있습니다.`,
+};
