+++ frontend/src/pages/mentoringCreate/utils/introduceValidator.ts
@@ -1,5 +1,4 @@
-import { ERROR_MESSAGE } from '../constants/errorMessage';
-import { INTRODUCE } from '../constants/introduce';
+import { ERROR_MESSAGE, INTRODUCE } from '../constants/introduce';
 
 export const introduceValidator = (introduction: string | null): string => {
   if (introduction === null) {
